+++ doc/source/changelog.rst
@@ -14,6 +14,7 @@ master: 1.7.0
 -------------
 
 * Upgrade to cppyy-cling 6.20.2
+* Fix crash when handing a by-value object to make_shared
 
 
 2020-03-15: 1.6.2